third_party/blink/web_tests/platform/generic/fast/multicol/vertical-rl/image-inside-nested-blocks-with-border-expected.txt

Test that getBoundingClientRect works as expected on a float in the second column - writing-mode is vertical-rl.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS result.left is 210
PASS result.top is 300
PASS successfullyParsed is true

TEST COMPLETE
